Create abc
This project runs a game called Bubble Pop. It is a game that allows you to shoot bubbles with one of 3 (or 6) colors. There is a hex grid of bubbles that you need to pop. Three consecutive balls of the same color will pop them all. The purpose of the game is to last as long as possible before the hex grid passes the line at the bottom.